feat & refactor: implemented cat command & and refactor file structure
