+++ handlers/files/index.js
@@ -1,0 +1,1 @@
+export { readFileHandler } from "./readFileHandler.js";
