+++ handlers/navigation/dedicateFolderHandler.js
