+++ handlers/navigation/readDirHandler.js
